Separando validacao de user em uma helper especifica
