+++ app/controllers/users_controller.rb
@@ -1,6 +1,6 @@
 require "rest-client"
 class UsersController < ApplicationController
-  include ValidationsHelper
+  include UserValidationHelper
   include UsersDoc
 
   skip_before_action :authenticate_request, only: [:create, :all]
