+++ app/helpers/validations_helper.rb
@@ -1,11 +1,5 @@
 module ValidationsHelper
-  def get_current_user
-    @current_user = AuthorizeApiRequest.call(request.headers).result
-  end
-
-  def get_project_user
-    @user = User.find(@project.user_id)
-  end
+  include UserValidationHelper
 
   def get_release_project
     @project = Project.find(@release.project_id)
@@ -29,10 +23,6 @@ module ValidationsHelper
 
   def get_retrospective_sprint
     @sprint = Sprint.find(@retrospective.sprint_id)
-  end
-
-  def get_user(id)
-    @user = User.find(id.to_i)
   end
 
   def get_project(id)
@@ -61,21 +51,6 @@ module ValidationsHelper
 
   def get_retrospective(id)
     @retrospective = Retrospective.find(id.to_i)
-  end
-
-  def validate_authorization
-    if @current_user.id == @user.id
-      return true
-    else
-      render json: { error: "Not Authorized" }, status: 401
-    end
-  end
-
-  def validate_user(user_id)
-    get_current_user
-    get_user(user_id)
-
-    validate_authorization
   end
 
   def validate_project(project_id)
